Pull the loadUsers into channel saga
