+++ src/store/channels-list/saga.test.ts
@@ -1,8 +1,8 @@
 import { expectSaga } from 'redux-saga-test-plan';
 import * as matchers from 'redux-saga-test-plan/matchers';
 
-import { fetchChannels } from './api';
-import { fetch, stopSyncChannels, unreadCountUpdated } from './saga';
+import { fetchChannels, fetchUsersByChannelId } from './api';
+import { channelIdPrefix, fetch, loadUsers, stopSyncChannels, unreadCountUpdated } from './saga';
 
 import { setStatus } from '.';
 import { rootReducer } from '..';
@@ -15,6 +15,17 @@ const MOCK_CHANNELS = [
 ];
 
 describe('channels list saga', () => {
+  const usersResponse = [
+    {
+      userId: 'the-first-id',
+      firstName: 'the first name',
+    },
+    {
+      userId: 'the-second-id',
+      firstName: 'the second name',
+    },
+  ];
+
   it('sets status to fetching', async () => {
     await expectSaga(fetch, { payload: '0x000000000000000000000000000000000000000A' })
       .put(setStatus(AsyncListStatus.Fetching))
@@ -125,12 +136,80 @@ describe('channels list saga', () => {
   });
 
   it('sets status to Stopped', async () => {
-    const id = '0x000000000000000000000000000000000000000A';
-
     const {
       storeState: { channelsList },
     } = await expectSaga(stopSyncChannels).withReducer(rootReducer).run();
 
     expect(channelsList.status).toBe(AsyncListStatus.Stopped);
+  });
+
+  it('load users', async () => {
+    const channelId = '0x000000000000000000000000000000000000000A';
+    await expectSaga(loadUsers, { payload: { channelId } })
+      .provide([
+        [
+          matchers.call.fn(fetchUsersByChannelId),
+          usersResponse,
+        ],
+      ])
+      .call(fetchUsersByChannelId, channelIdPrefix + channelId)
+      .run();
+  });
+
+  it('adds users ids to channels state', async () => {
+    const channelId = 'channel-id';
+
+    const initialState = {
+      normalized: {
+        channels: {
+          [channelId]: {
+            id: channelId,
+            users: [
+              'old-user-id',
+            ],
+          },
+        },
+      },
+    };
+
+    const {
+      storeState: {
+        normalized: { channels },
+      },
+    } = await expectSaga(loadUsers, { payload: { channelId } })
+      .withReducer(rootReducer, initialState as any)
+      .provide([
+        [
+          matchers.call.fn(fetchUsersByChannelId),
+          usersResponse,
+        ],
+      ])
+      .run();
+
+    expect(channels[channelId].users).toStrictEqual([
+      'the-first-id',
+      'the-second-id',
+    ]);
+  });
+
+  it('adds users to normalized state', async () => {
+    const {
+      storeState: {
+        normalized: { users },
+      },
+    } = await expectSaga(loadUsers, { payload: { channelId: '0x000000000000000000000000000000000000000A' } })
+      .provide([
+        [
+          matchers.call.fn(fetchUsersByChannelId),
+          usersResponse,
+        ],
+      ])
+      .withReducer(rootReducer)
+      .run();
+
+    expect(users).toMatchObject({
+      'the-first-id': { id: 'the-first-id', firstName: 'the first name' },
+      'the-second-id': { id: 'the-second-id', firstName: 'the second name' },
+    });
   });
 });
