--- src/store/users/api.ts
@@ -1,7 +1,0 @@
-import { get } from '../../lib/api/rest';
-
-export async function fetchUsersByChannelId(channelId: string): Promise<any> {
-  const response = await get<any>(`/chatChannels/${channelId}/members`);
-
-  return response.body;
-}
